move and refactor features.py to utils and refactor utils.py --> pc_utils.py
